[All] Move backlog upstream

+++ src/Mono.WebServer.Apache/main.cs
@@ -175,7 +175,7 @@ namespace Mono.WebServer.Apache {
 			}
 
 			try {
-				if (server.Start (!configurationManager.NonStop, configurationManager.Backlog) == false)
+				if (server.Start (!configurationManager.NonStop, (int)configurationManager.Backlog) == false)
 					return 2;
 
 				if (!configurationManager.NonStop) {
